Mask off support of events for the different releases
Not all events are supported by every release of the Switchtec
firmware.  To solve this, add a rel_mask field to the event_regs
table which is a bitmask of all supported releases for a given event.

Check this mask as part of validating and event_ctl IOCTL and when
clearing all the event bits in the ISR.

Signed-off-by: Logan Gunthorpe <logang@deltatee.com>