Mask off support of events for the different gens
Not all events are supported by every generation of hardware. To solve
this, add a gen_mask field to the event_regs table which is a bitmask
of all supported generations for a given event.

Check this mask as part of validating and event_ctl IOCTL and when
clearing all the event bits in the ISR.

Signed-off-by: Logan Gunthorpe <logang@deltatee.com>