Introduce Event List Version
GEN4 introduces a version number for the event list so the driver
can tell which events are applicable. For GEN3, we have to emulate this
based on the firmware release version.

The GEN4 version starts at 1 and will increments from there. We put the GEN3
versions at 25 and up (seeing we will want to fit these in a 32bit
bitmask). This gives us 24 new versions and if/when we run out of space
we can easily relocate the the GEN3 versions to a higher number (seeing
these numbers are private to the kernel).

Signed-off-by: Logan Gunthorpe <logang@deltatee.com>